+++ client/common/utils/anti-csrf-token.js
