+++ client/common/utils/auth-check.js
