+++ client/common/utils/default-xhr-headers.js
