+++ client/common/utils/is-logged-in.js
