+++ client/common/utils/js-emoji.js
