+++ client/common/utils/load-initial-data.js
