+++ client/common/utils/string-util.js
