+++ client/common/utils/validate-emoji.js
