+++ client/index.js
@@ -20,8 +20,8 @@ import Privacy from './meta/components/privacy';
 import Terms from './meta/components/terms';
 import SignIn from './meta/components/sign-in';
 import store from './state/store';
-import generateAuthCheck from './common/services/auth-check';
-import {getYearMonthStringFromDate} from './transactions/services/format-date';
+import generateAuthCheck from './common/utils/auth-check';
+import {getYearMonthStringFromDate} from './transactions/utils/format-date';
 
 if (process.NODE_ENV !== 'production') {
   window.store = store;
