+++ client/transactions/components/content.js
@@ -11,9 +11,9 @@ import EmptyTransactions from './empty-transactions';
 import ErrorRetrieving from '../../common/components/error-retrieving';
 import LoadingResourceList from '../../common/components/loading-resource-list';
 import NotFound from '../../common/components/not-found';
-import {getYearMonthStringFromDate} from '../services/format-date';
-import validateTransactionDate from '../services/validate-transaction-date';
-import monthDiff from '../services/month-diff';
+import {getYearMonthStringFromDate} from '../utils/format-date';
+import validateTransactionDate from '../utils/validate-transaction-date';
+import monthDiff from '../utils/month-diff';
 
 export class Content extends Component {
   render() {
