+++ test/unit/client/categories/components/category-list-item.js
@@ -1,6 +1,6 @@
 import CategoryListItem from '../../../../../client/categories/components/category-list-item';
-import emoji from '../../../../../client/common/services/js-emoji';
-import generateWrapperGenerator from '../../../../services/generate-wrapper-generator';
+import emoji from '../../../../../client/common/utils/js-emoji';
+import generateWrapperGenerator from '../../../../utils/generate-wrapper-generator';
 
 describe('CategoryListItem', function() {
   beforeEach(() => {
