+++ test/unit/client/categories/components/modify-category-modal.js
@@ -1,7 +1,7 @@
 import {
   ModifyCategoryModal, __GetDependency__
 } from '../../../../../client/categories/components/modify-category-modal';
-import generateWrapperGenerator from '../../../../services/generate-wrapper-generator';
+import generateWrapperGenerator from '../../../../utils/generate-wrapper-generator';
 
 const validate = __GetDependency__('validate');
