+++ test/unit/client/common/services/load-initial-data.js
@@ -1,4 +1,4 @@
-import loadInitialData from '../../../../../client/common/services/load-initial-data';
+import loadInitialData from '../../../../../client/common/utils/load-initial-data';
 
 describe('loadInitialData', function() {
   it('should return {} when there is no initial data element', () => {
