+++ test/unit/client/transactions/services/validate-transaction-date.js
@@ -1,4 +1,4 @@
-import validateTransactionDate from '../../../../../client/transactions/services/validate-transaction-date';
+import validateTransactionDate from '../../../../../client/transactions/utils/validate-transaction-date';
 
 describe('validateTransactionDate', function() {
   it('should return false for undefined and null', () => {
